+++ CHANGELOG.md
@@ -28,6 +28,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Minder shotmeldingen wanneer je wapen een suppressor heeft
 - Het restaurant leftover buy menu sluit niet meer wanneer je iets koopt
 - Enkele crafting requirements verlaagd
+- 'Gemarkeerd voor inbeslagname' staat nu in een kleur in je inventory om het duidelijker te maken
 
 ### Fixed
 - De driver bij een ATM robbery hoeft niet langer uit te stappen voordat hij de ATM van de muur kan trekken